requirements.txt
neo4j==5.20.0
python-dotenv==1.0.1
streamlit==1.33.0
GraphDataScience==1.10